requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ccxt==3.0.24
discord.py==2.3.1
numpy==1.23.0
pandas==2.0.3
schedule==1.2.0
PyPortfolioOpt